scripts/requirements.txt
bitsandbytes
transformers
peft
accelerate
trl
pandas
Pillow
wandb
sentencepiece
evaluate

